Add files via upload
Completely implemented the frontend part as part of the creation of the Edchart project.
The project is an aggregator of courses and allows the user to perform a comfortable search for online schools and courses, get acquainted with their educational programs and choose the best one for themselves. The platform also allows you to get acquainted with the rating of schools and reviews on them and leave a review yourself.
Designed and created new functionality, participated in the code review process, assessed the complexity of development and participated in planning.
I added new functionality, fixed bugs, did code reviews and participated in studying the work of competitors. Optimized the code, participated in discussions about adding something new to the project. Developed the client side of the site on React, React query, worked with the API. From scratch, I implemented authorization, registration, in conjunction with the backend, wrote JS code in conjunction with React for the entire project. I also implemented from scratch and connected with the backend in the feedback form, the slider, on the pages: school rankings, school reviews, courses (which includes filters and course cards), as well as the main page. Worked with Docker.